% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/char2number.R
\name{char2number}
\alias{char2number}
\title{Converts a string to a number by character substitution}
\usage{
char2number(init.string, chars.list = "NULL", num.list = "NULL",
  mod.machine.int = FALSE)
}
\arguments{
\item{init.string}{character string to be converted}

\item{chars.list}{list of characters to be replaced.  Default is c('a',...,'z','A',...,'Z')}

\item{num.list}{list of numbers to replace characters in chars.list.  Default is seq(from = 1, to = 52, by = 1)}

\item{mod.machine.int}{bool to return the full number (default = FALSE) or the full number modulus the integer range of the machine (set to TRUE).}
}
\value{
number
}
\description{
Currently only supports a-zA-z and removes spaces
}
\examples{
char2number('a')
char2number("Hello World")
}

